[XmlTreeView] more straightforward node access from scripts
Instead of tracking & exposing the currently selected node on the
TreeView simply expose a getter which returns a node proxy for the given
tree item id.

Not entirely sure how this affects runtime perf & mem footprint but it
greatly simplifies the implementation on the c++ side and makes usage of
it from scripts much more straightforward (as we now don't have to think
about strange lifetime rules). This is also much more amenable to adding
"editability".